ajustes-formulario: move componentes para a pasta /src.
