+++ src/screens/Home/index.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import {ImageBackground, Text, View} from 'react-native';
-const backgroundImage = require('../../assets/background.png');
-import globalStyles from '@shared/globalStyles';
-
-import NavigationButton from '@components/NavigationButton';
+const backgroundImage = require('../../../assets/background.png');
+import NavigationButton from '../../components/NavigationButton';
+import globalStyles from '../../shared/globalStyles';
 
 function Home({}) {
   return (
