+++ src/screens/ListCard/styles.ts
