+++ src/shared/globalStyles.ts
