+++ tsconfig.json
@@ -4,11 +4,11 @@
     "baseUrl": ".",
     "paths": {
       "tests": ["tests/*"],
-      "@screens/*": ["screens/*"],
-      "@components/*": ["components/*"],
-      "@models/*": ["models/*"],
-      "@api/*": ["api/*"],
-      "@shared/*": ["shared/*"],
+      "@screens/*": ["src/screens/*"],
+      "@components/*": ["src/components/*"],
+      "@models/*": ["src/models/*"],
+      "@api/*": ["src/api/*"],
+      "@shared/*": ["src/shared/*"],
     },
   }
 }
